+++ core/server/api_container/server/service_network/default_service_network.go
@@ -605,25 +605,25 @@ func (network *DefaultServiceNetwork) GetServiceIDs() map[service.ServiceID]bool
 	return serviceIDs
 }
 
-func (network *DefaultServiceNetwork) CopyFilesFromService(ctx context.Context, serviceId service.ServiceID, srcPath string) (enclave_data_directory.FilesArtifactUUID, error) {
-	filesArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+func (network *DefaultServiceNetwork) CopyFilesFromService(ctx context.Context, serviceId service.ServiceID, srcPath string) (enclave_data_directory.FilesArtifactID, error) {
+	filesArtifactId, err := enclave_data_directory.NewFilesArtifactID()
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in creating a files artifact uuid to copy the files to")
 	}
 
-	err = network.copyFilesFromServiceToTargetArtifactUUIDUnlocked(ctx, serviceId, srcPath, filesArtifactUuid)
+	err = network.copyFilesFromServiceToTargetArtifactIdUnlocked(ctx, serviceId, srcPath, filesArtifactId)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in copying files over to disk")
 	}
-	return filesArtifactUuid, nil
+	return filesArtifactId, nil
 }
 
-func (network *DefaultServiceNetwork) CopyFilesFromServiceToTargetArtifactUUID(ctx context.Context, serviceId service.ServiceID, srcPath string, filesArtifactUuid enclave_data_directory.FilesArtifactUUID) (enclave_data_directory.FilesArtifactUUID, error) {
-	err := network.copyFilesFromServiceToTargetArtifactUUIDUnlocked(ctx, serviceId, srcPath, filesArtifactUuid)
+func (network *DefaultServiceNetwork) CopyFilesFromServiceToTargetArtifactUUID(ctx context.Context, serviceId service.ServiceID, srcPath string, filesArtifactId enclave_data_directory.FilesArtifactID) (enclave_data_directory.FilesArtifactID, error) {
+	err := network.copyFilesFromServiceToTargetArtifactIdUnlocked(ctx, serviceId, srcPath, filesArtifactId)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in copying files over to disk")
 	}
-	return filesArtifactUuid, nil
+	return filesArtifactId, nil
 }
 
 func (network *DefaultServiceNetwork) GetIPAddressForService(serviceID service.ServiceID) (net.IP, bool) {
@@ -636,50 +636,50 @@ func (network *DefaultServiceNetwork) GetIPAddressForService(serviceID service.S
 	return registration.GetPrivateIP(), true
 }
 
-func (network *DefaultServiceNetwork) RenderTemplates(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData) (enclave_data_directory.FilesArtifactUUID, error) {
+func (network *DefaultServiceNetwork) RenderTemplates(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData) (enclave_data_directory.FilesArtifactID, error) {
 	// calling mutex just in case, even though we don't change or read from the network, just get the file store
 	network.mutex.Lock()
 	defer network.mutex.Unlock()
 
-	filesArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+	filesArtifactId, err := enclave_data_directory.NewFilesArtifactID()
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in creating a files artifact uuid to render the templates to")
 	}
-	err = network.renderTemplatesToTargetArtifactUUIDUnlocked(templatesAndDataByDestinationRelFilepath, filesArtifactUuid)
+	err = network.renderTemplatesToTargetArtifactIDUnlocked(templatesAndDataByDestinationRelFilepath, filesArtifactId)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in rendering templates to disk")
 	}
-	return filesArtifactUuid, nil
+	return filesArtifactId, nil
 }
 
-func (network *DefaultServiceNetwork) RenderTemplatesToTargetFilesArtifactUUID(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, filesArtifactUuid enclave_data_directory.FilesArtifactUUID) (enclave_data_directory.FilesArtifactUUID, error) {
+func (network *DefaultServiceNetwork) RenderTemplatesToTargetFilesArtifactUUID(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, filesArtifactUuid enclave_data_directory.FilesArtifactID) (enclave_data_directory.FilesArtifactID, error) {
 	// calling mutex just in case, even though we don't change or read from the network, just get the file store
 	network.mutex.Lock()
 	defer network.mutex.Unlock()
 
-	err := network.renderTemplatesToTargetArtifactUUIDUnlocked(templatesAndDataByDestinationRelFilepath, filesArtifactUuid)
+	err := network.renderTemplatesToTargetArtifactIDUnlocked(templatesAndDataByDestinationRelFilepath, filesArtifactUuid)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in rendering templates to disk")
 	}
 	return filesArtifactUuid, nil
 }
 
-func (network *DefaultServiceNetwork) UploadFilesArtifact(data []byte) (enclave_data_directory.FilesArtifactUUID, error) {
-	filesArtifactUuid, err := enclave_data_directory.NewFilesArtifactUUID()
+func (network *DefaultServiceNetwork) UploadFilesArtifact(data []byte) (enclave_data_directory.FilesArtifactID, error) {
+	filesArtifactId, err := enclave_data_directory.NewFilesArtifactID()
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in creating a files artifact uuid to upload the files to")
 	}
 
-	err = network.uploadFilesArtifactToTargetArtifactUUIDUnlocked(data, filesArtifactUuid)
+	err = network.uploadFilesArtifactToTargetArtifactIDUnlocked(data, filesArtifactId)
 	if err != nil {
 		return "", stacktrace.Propagate(err, "There was an error in uploading the files")
 	}
 
-	return filesArtifactUuid, nil
+	return filesArtifactId, nil
 }
 
-func (network *DefaultServiceNetwork) UploadFilesArtifactToTargetArtifactUUID(data []byte, targetFilesArtifactUuid enclave_data_directory.FilesArtifactUUID) error {
-	err := network.uploadFilesArtifactToTargetArtifactUUIDUnlocked(data, targetFilesArtifactUuid)
+func (network *DefaultServiceNetwork) UploadFilesArtifactToTargetArtifactID(data []byte, targetFilesArtifactId enclave_data_directory.FilesArtifactID) error {
+	err := network.uploadFilesArtifactToTargetArtifactIDUnlocked(data, targetFilesArtifactId)
 	if err != nil {
 		return stacktrace.Propagate(err, "There was an error in uploading the files")
 	}
@@ -851,7 +851,7 @@ func (network *DefaultServiceNetwork) startServices(
 }
 
 // This method is not thread safe. Only call this from a method where there is a mutex lock on the network.
-func (network *DefaultServiceNetwork) copyFilesFromServiceToTargetArtifactUUIDUnlocked(ctx context.Context, serviceId service.ServiceID, srcPath string, filesArtifactUuId enclave_data_directory.FilesArtifactUUID) error {
+func (network *DefaultServiceNetwork) copyFilesFromServiceToTargetArtifactIdUnlocked(ctx context.Context, serviceId service.ServiceID, srcPath string, filesArtifactId enclave_data_directory.FilesArtifactID) error {
 	serviceObj, found := network.registeredServiceInfo[serviceId]
 	if !found {
 		return stacktrace.NewError("Cannot copy files from service '%v' because it does not exist in the network", serviceId)
@@ -872,7 +872,7 @@ func (network *DefaultServiceNetwork) copyFilesFromServiceToTargetArtifactUUIDUn
 		defer pipeReader.Close()
 
 		//And finally pass it the .tgz file to the artifact file store
-		storeFileErr := store.StoreFileToArtifactUUID(pipeReader, filesArtifactUuId)
+		storeFileErr := store.StoreFileToArtifactUUID(pipeReader, filesArtifactId)
 		storeFilesArtifactResultChan <- storeFilesArtifactResult{
 			err: storeFileErr,
 		}
@@ -974,7 +974,7 @@ func (network *DefaultServiceNetwork) createSidecarAndAddToMap(ctx context.Conte
 }
 
 // This method is not thread safe. Only call this from a method where there is a mutex lock on the network.
-func (network *DefaultServiceNetwork) renderTemplatesToTargetArtifactUUIDUnlocked(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, filesArtifactUuid enclave_data_directory.FilesArtifactUUID) error {
+func (network *DefaultServiceNetwork) renderTemplatesToTargetArtifactIDUnlocked(templatesAndDataByDestinationRelFilepath map[string]*kurtosis_core_rpc_api_bindings.RenderTemplatesToFilesArtifactArgs_TemplateAndData, filesArtifactId enclave_data_directory.FilesArtifactID) error {
 	tempDirForRenderedTemplates, err := os.MkdirTemp("", tempDirForRenderedTemplatesPrefix)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred while creating a temp dir for rendered templates '%v'", tempDirForRenderedTemplates)
@@ -1014,15 +1014,15 @@ func (network *DefaultServiceNetwork) renderTemplatesToTargetArtifactUUIDUnlocke
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred while getting files artifact store")
 	}
-	err = store.StoreFileToArtifactUUID(bytes.NewReader(compressedFile), filesArtifactUuid)
+	err = store.StoreFileToArtifactUUID(bytes.NewReader(compressedFile), filesArtifactId)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred while storing the file '%v' in the files artifact store", compressedFile)
 	}
 	shouldDeleteFilesArtifact := true
 	defer func() {
 		if shouldDeleteFilesArtifact {
-			if err = store.RemoveFile(filesArtifactUuid); err != nil {
-				logrus.Errorf("We tried to clean up the files artifact '%v' we had stored but failed:\n%v", filesArtifactUuid, err)
+			if err = store.RemoveFile(filesArtifactId); err != nil {
+				logrus.Errorf("We tried to clean up the files artifact '%v' we had stored but failed:\n%v", filesArtifactId, err)
 			}
 		}
 	}()
@@ -1032,7 +1032,7 @@ func (network *DefaultServiceNetwork) renderTemplatesToTargetArtifactUUIDUnlocke
 }
 
 // This method is not thread safe. Only call this from a method where there is a mutex lock on the network.
-func (network *DefaultServiceNetwork) uploadFilesArtifactToTargetArtifactUUIDUnlocked(data []byte, targetFilesArtifactUuid enclave_data_directory.FilesArtifactUUID) error {
+func (network *DefaultServiceNetwork) uploadFilesArtifactToTargetArtifactIDUnlocked(data []byte, targetFilesArtifactId enclave_data_directory.FilesArtifactID) error {
 	reader := bytes.NewReader(data)
 
 	filesArtifactStore, err := network.enclaveDataDir.GetFilesArtifactStore()
@@ -1040,7 +1040,7 @@ func (network *DefaultServiceNetwork) uploadFilesArtifactToTargetArtifactUUIDUnl
 		return stacktrace.Propagate(err, "An error occurred while getting files artifact store")
 	}
 
-	err = filesArtifactStore.StoreFileToArtifactUUID(reader, targetFilesArtifactUuid)
+	err = filesArtifactStore.StoreFileToArtifactUUID(reader, targetFilesArtifactId)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred while trying to store files.")
 	}